% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Horizontal.R
\name{AddShift}
\alias{AddShift}
\title{Add or Shift a Phoneme}
\usage{
AddShift(P, targetLanguage, languages, local, phonemeRelatedness, index)
}
\arguments{
\item{P}{A list of parameters.}

\item{targetLanguage}{The target language to be modified if possible.}

\item{languages}{All languages}

\item{local}{The local territories data structure.}

\item{index}{The target territory whose language may change.}

\item{phonemeProbab}{The probability of gaining each phoneme in the population.}
}
\description{
Allows a language to either gain a new phoneme or shift an existing phoneme to better match another population's phoneme iventory.
}
\keyword{Horizontal}
