% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{BeringStraitPlot}
\alias{BeringStraitPlot}
\title{Bering Straight Plot}
\usage{
BeringStraitPlot(P, Data, colors = NA)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}

\item{colors}{A vector of colors of length equal to the number of seed populations.}
}
\description{
Creates a plot that shows the Bring strait boundaries.
}
\keyword{Plotting}
