% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenLanguage.R
\name{GetASplitShiftJoinMut}
\alias{GetASplitShiftJoinMut}
\title{Get A Split Shift Join Mutation}
\usage{
GetASplitShiftJoinMut(phonemes, phonemeRelatedness, unusable = NULL,
  type)
}
\arguments{
\item{phonemes}{Phonemes that can be mutated.}

\item{phonemeRelatedness}{The phoneme relatedness list.}

\item{unusable}{Phonemes that cannot be used to obtain the correct type of mutation.}

\item{type}{Which kind of mutation to create: Split, Join, or Shift.}
}
\description{
Recursively calls itself until a phoneme is found that can be used the generate the mutation type of interest.  Returns Null if no phoneme can mutate appropriately.
}
\keyword{Language}
