% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Barriers.R
\name{GetBering}
\alias{GetBering}
\title{Get Bering positions}
\usage{
GetBering(P)
}
\arguments{
\item{P}{A list of parameters.}
}
\description{
returns a list of the points for the berring strait
}
\keyword{Barriers}
