% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{GetBeringCoords}
\alias{GetBeringCoords}
\title{Get Bering Strait Coordinates}
\usage{
GetBeringCoords(P)
}
\description{
Returns the hardcoded locations of the Bering Strait boundaries.
}
\keyword{Plotting}
