% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{GetGroups}
\alias{GetGroups}
\title{Get Groups}
\usage{
GetGroups(P, Data)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}
}
\description{
Returns the territories descended from each seed.  Includes detailed ancestory data.  Ony works when Uproot and Death are FALSE.
}
\keyword{Plotting}
