% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Phonemes.R
\name{GetRealPhonemeData}
\alias{GetRealPhonemeData}
\title{Get Real Phoneme Data}
\usage{
GetRealPhonemeData(nPhoneme, actual, vowel = FALSE)
}
\arguments{
\item{nPhoneme}{The number of phonemes (vowels or consonants).}

\item{actual}{Whether the data is Real (TRUE) or RealMimic (FALSE).}

\item{vowel}{If true load the vowel data, otherwise load the cosonant data.}
}
\description{
Uses the real Phoneme data from Creanza..... UPDATE THIS!!!! to determine the phoneme probabilities.
}
\keyword{Phonemes}
