% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{GetXYCoords}
\alias{GetXYCoords}
\title{Get XY Coordinates}
\usage{
GetXYCoords(P, territories)
}
\arguments{
\item{P}{A list of parameters.}

\item{territories}{A vector of territory indicies to convert into X,Y coordinates.}
}
\description{
COnverts territory numbers into X,Y corrdinates.
}
\keyword{Plotting}
