% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Migration.R
\name{NextWave}
\alias{NextWave}
\title{Next Wave}
\usage{
NextWave(P, S, i)
}
\arguments{
\item{P}{A list of parameters.}

\item{S}{A list of data structures.}

\item{i}{The number of the next wave.}
}
\description{
Adds the seed data for the next wave to the population and language dataframes.
}
\keyword{SimParam}
