% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Barriers.R
\name{RemoveHorizontalConnections}
\alias{RemoveHorizontalConnections}
\title{Remove Horizontal Connections}
\usage{
RemoveHorizontalConnections(R, index, firstStep, right = TRUE,
  left = TRUE)
}
\arguments{
\item{R}{The number of rows in the population matrix.}

\item{index}{The target territory.}

\item{firstStep}{The local directions created by OneStepDirections().}

\item{right}{Whether to remove the right diagonal.}

\item{left}{Whether to remove the left diagonal.}
}
\description{
Affects local territories below/South (and perhaps to the Southeast and Southwest) the target territory (index) and above/North (perhaps Northwest and Northeast) of index +1.
}
\keyword{Barriers}
