% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{UpdateStructuresMove}
\alias{UpdateStructuresMove}
\title{Update Structures Move}
\usage{
UpdateStructuresMove(S, move, former)
}
\arguments{
\item{S}{A list of data structures.}

\item{move}{The indicies of territories}

\item{former}{The indicies of territories}
}
\description{
Copies population data from one territory to another when the entire population migrates and then erases the original data.
}
\keyword{Population_Dynamics}
